# openapi.yml
openapi: 3.1.0
info:
  title: Speaker API
  version: 0.0.1
servers:
  - url: https://speaker-be.egor-koldasov.dev
    description: Production server
components:
  schemas:
    Word:
      description: >
        Example:
        {
          "originalWord": "Esperanzamos",
          "neutralForm": "Esperanzar",
          "pronounciation": "/es.pe.ɾanˈθa.mos/ (Spain), /es.pe.ɾanˈsa.mos/ (Latin America)",
          "translationEnglish": "To fill with hope, to become hopeful, to start to hope, to hold onto hope.",
          "synonyms": [
            "Ilusionar",
            "tener esperanza",
            "confiar"
          ],
          "definitionOriginal": "Acción de hacerse llevar por la esperanza o comenzar a tener esperanza frente a una situación, a pesar de las circunstancias que pueden sugerir lo contrario.",
          "definitionEnglish": "The action of being guided by or starting to have hope in a situation, despite circumstances that might suggest the contrary.",
          "origin": "The verb \"esperanzar\" is derived from the noun \"esperanza,\" meaning hope. The suffix \"-ar\" indicates that it's an infinitive verb. In this context, \"esperanzamos\" is the first person plural of the present indicative, meaning \"we hope\" or \"we become hopeful.\" The transition from the neutral form \"esperanzar\" to \"esperanzamos\" changes the verb from its infinitive form to a specific tense and subject, indicating the action of hoping being performed by us in the present.",
          "examples": [
            {
              "original": "Y nos esperanzamos contra toda lógica, contra todo lo que nos dice la experiencia.",
              "english": "And we become hopeful against all logic, against everything our experience tells us."
            },
            {
              "original": "Cada vez que sale el sol, nos esperanzamos pensando que será un mejor día.",
              "english": "Every time the sun rises, we become hopeful thinking it will be a better day."
            },
            {
              "original": "A pesar de las dificultades, nos esperanzamos al ver la solidaridad de la gente.",
              "english": "Despite the difficulties, we become hopeful upon seeing people's solidarity."
            }
          ]
        }

      type: object
      required:
        - originalWord
        - neutralForm
        - pronounciation
        - translationEnglish
        - synonyms
        - definitionOriginal
        - definitionEnglish
        - origin
        - examples
      properties:
        originalWord:
          type: string
          description: The original word given, in the exact same grammatic form, capitalized
        neutralForm:
          type: string
          description: The word in a neutral grammatic form
        pronounciation:
          type: string
          description: A pronunciation of the original word given
        translationEnglish:
          type: string
          description: An extensive translation to English, the more words the better
        synonyms:
          description: Common synonyms in the original language
          type: array
          items:
            type: string
        definitionOriginal:
          description: An extensive definition in the original language
          type: string
        definitionEnglish:
          description: An extensive definition in English
          type: string
        origin:
          description: The root parts of the word and the origin in English. If the original form from Part 1 is different from the neutral grammatic form from Part 2, explain that difference including all the details.
          type: string
        examples:
          description:
            Three sentence examples of the usage of the original word in the same grammatic form followed by an English translation.
            The sentence and the translation should be separated by one new line,
            while the examples themselves should be separated by three new lines.
            If there was a context from which that word was taken, include a phrase from that context in examples,
            replacing the first example.
          type: array
          items:
            type: object
            properties:
              original:
                type: string
              english:
                type: string
  responses:
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
paths:
  /word:
    get:
      operationId: getWords
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
          required: false
          description: Number of words to return, default is 100
      responses:
        '200':
          description: A list of words
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Word'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      operationId: createWord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Word'
      responses:
        '201':
          description: Word created
        '500':
          $ref: '#/components/responses/ServerError'
  /csv:
    get:
      operationId: exportCSV
      responses:
        '200':
          description: Export words as CSV
          content:
            text/csv:
              schema:
                type: string
        '500':
          $ref: '#/components/responses/ServerError'